map popup now displays picture of property
