+++ Backend/config/database.yml
@@ -46,7 +46,7 @@ development:
 
   # The TCP port the server listens on. Defaults to 5432.
   # If your server runs on a different port number, change accordingly.
-  port: 5432
+  port: 5433
   
 
   # Schema search path. The server defaults to $user,public
