+++ Frontend/src/api/feedbackEndpoints.js
@@ -1,6 +1,6 @@
 import axios from 'axios';
 
-const BASE_URL = 'http://localhost:3000';
+const BASE_URL = 'http://localhost:3001';
 
 const feedback_endpoints = {
   GET_FEEDBACK: `${BASE_URL}/show`,
