+++ Frontend/src/components/PropertyCard.jsx
@@ -1,16 +1,8 @@
 import React from "react";
 import ApplyButton from "./Button Components/ApplyButton";
+import { getImageURL } from "../helpers/getImageURL";
 
 function PropertyCard({ listing, onClick, User, isTenant }) {
-  // Helper function to dynamically build Unsplash image URL
-  const getImageURL = (type) => {
-    const fallback = "/images/fallback.jpg";
-    if (typeof type === "string" && type.trim() !== "") {
-      const formatted = type.toLowerCase().trim().replace(/_/g, " ");
-      return `/images/${formatted}.jpg`;
-    }
-    return fallback;
-  };
 
   return (
     <div
