+++ Frontend/src/components/ViewPropertiesDashboard/LeafletMapContainer.jsx
@@ -1,13 +1,17 @@
 import React from 'react'
 import { MapContainer, TileLayer, useMap, Marker, Popup } from 'react-leaflet'
+import { getImageURL } from '../../helpers/getImageURL';
 import "leaflet/dist/leaflet.css";
+
 
 import { LocationPin } from '../../JS/LocationPin';
 import { useNavigate } from 'react-router-dom';
-import { useState, useEffect } from 'react'
+
+function LeafletMapContainer({listings, geoLocation}) {
 
 
-function LeafletMapContainer({listings}) {
+
+    console.log("geoLocation", geoLocation)
     
     const navigateToProperty = useNavigate()
 
@@ -15,8 +19,8 @@ function LeafletMapContainer({listings}) {
         <MapContainer
             className="Leaflet_map"
             style={{ height: "800px" }}
-            center={[51.505, -0.09]}
-            zoom={2.25}
+            center={[53, -110]}
+            zoom={3}
             scrollWheelZoom={true}
         >
             <TileLayer
@@ -37,7 +41,7 @@ function LeafletMapContainer({listings}) {
                                 <img
                                     cursor="pointer"
                                     onClick={() => navigateToProperty('property-page/' + listing.id)}
-                                    src={listing.image}
+                                    src={getImageURL(listing.property_types)}
                                     alt={listing.title}
                                     style={{ width: "100px", height: "100px" }}
                                 />
@@ -54,6 +58,9 @@ function LeafletMapContainer({listings}) {
                 }
                 return null; // Return null if latitude or longitude is missing
             })}
+            <Marker 
+                position={[geoLocation.latitude, geoLocation.longitude]}
+            />
         </MapContainer>
     );
 }
